refactor(email): move email routes to /api/zero/email/ namespace
Move all email integration routes from infra-layer paths to the zero
app layer namespace for consistency with Slack (/api/zero/slack/).

- Move inbound webhook to /api/zero/email/inbound
- Move trigger/reply/schedule callbacks to /api/zero/email/callbacks/*
- Add re-export redirects at old paths for backward compatibility
- Update callback URL generation in handlers and schedule service
- Move tests to new locations and update URL references

Part of #6226. Resend dashboard webhook URL needs manual update after deploy.
