+++ turbo/apps/web/app/api/zero/email/callbacks/schedule/__tests__/route.test.ts
@@ -48,7 +48,7 @@ function createCallbackRequest(
     : computeHmacSignature(bodyString, secret, timestamp);
 
   return createTestRequest(
-    "http://localhost/api/internal/callbacks/email/schedule",
+    "http://localhost/api/zero/email/callbacks/schedule",
     {
       method: "POST",
       headers: {
@@ -61,7 +61,7 @@ function createCallbackRequest(
   );
 }
 
-describe("POST /api/internal/callbacks/email/schedule", () => {
+describe("POST /api/zero/email/callbacks/schedule", () => {
   beforeEach(() => {
     context.setupMocks();
     mockResend.emails.send.mockClear();
@@ -86,7 +86,7 @@ describe("POST /api/internal/callbacks/email/schedule", () => {
 
       const { secret } = await createTestCallback({
         runId,
-        url: "http://localhost/api/internal/callbacks/email/schedule",
+        url: "http://localhost/api/zero/email/callbacks/schedule",
         payload: { ...payload },
       });
 
@@ -118,7 +118,7 @@ describe("POST /api/internal/callbacks/email/schedule", () => {
 
       const { secret } = await createTestCallback({
         runId,
-        url: "http://localhost/api/internal/callbacks/email/schedule",
+        url: "http://localhost/api/zero/email/callbacks/schedule",
         payload: { ...payload },
       });
 
@@ -152,7 +152,7 @@ describe("POST /api/internal/callbacks/email/schedule", () => {
 
       const { secret } = await createTestCallback({
         runId,
-        url: "http://localhost/api/internal/callbacks/email/schedule",
+        url: "http://localhost/api/zero/email/callbacks/schedule",
         payload: { ...payload },
       });
 
@@ -191,7 +191,7 @@ describe("POST /api/internal/callbacks/email/schedule", () => {
 
       const { secret } = await createTestCallback({
         runId,
-        url: "http://localhost/api/internal/callbacks/email/schedule",
+        url: "http://localhost/api/zero/email/callbacks/schedule",
         payload: { ...payload },
       });
 
@@ -234,7 +234,7 @@ describe("POST /api/internal/callbacks/email/schedule", () => {
 
       const { secret } = await createTestCallback({
         runId,
-        url: "http://localhost/api/internal/callbacks/email/schedule",
+        url: "http://localhost/api/zero/email/callbacks/schedule",
         payload: { ...payload },
       });
